+++ lifelines/utils/__init__.py
@@ -1352,4 +1352,7 @@ def format_floats(decimals):
     return lambda f: "{:4.{prec}f}".format(f, prec=decimals)
 
 
+def dataframe_interpolate_at_times(df, times):
+    return df.reindex(df.index.union(_to_array(times))).interpolate("index").loc[times].squeeze()
+
 string_justify = lambda width: lambda s: s.rjust(width, " ")