+++ tests/test_statistics.py
@@ -181,11 +181,11 @@ def test_pairwise_allows_dataframes_and_gives_correct_counts():
 
 def test_log_rank_returns_None_if_equal_arrays():
     T = np.random.exponential(5, size=200)
-    result = stats.logrank_test(T, T, alpha=0.95)
+    result = stats.logrank_test(T, T)
     assert result.p_value > 0.05
 
     C = np.random.binomial(1, 0.8, size=200)
-    result = stats.logrank_test(T, T, C, C, alpha=0.95)
+    result = stats.logrank_test(T, T, C, C)
     assert result.p_value > 0.05
 
 
@@ -195,12 +195,12 @@ def test_multivariate_log_rank_is_identital_to_log_rank_for_n_equals_2():
     T2 = np.random.exponential(5, size=N)
     C1 = np.random.binomial(1, 0.9, size=N)
     C2 = np.random.binomial(1, 0.9, size=N)
-    result = stats.logrank_test(T1, T2, C1, C2, alpha=0.95)
+    result = stats.logrank_test(T1, T2, C1, C2)
 
     T = np.r_[T1, T2]
     C = np.r_[C1, C2]
     G = np.array([1] * 200 + [2] * 200)
-    result_m = stats.multivariate_logrank_test(T, G, C, alpha=0.95)
+    result_m = stats.multivariate_logrank_test(T, G, C)
     assert result.p_value == result_m.p_value
 
 
@@ -222,13 +222,6 @@ def test_StatisticalResult_can_be_added():
     assert sr.summary.shape[0] == 4
     assert sr.summary.index.tolist() == ["1", "2", "3", "4"]
     assert "kw3" in sr._kwargs
-
-
-def test_valueerror_is_raised_if_alpha_out_of_bounds():
-    data1 = np.random.exponential(5, size=(20, 1))
-    data2 = np.random.exponential(1, size=(20, 1))
-    with pytest.raises(ValueError):
-        stats.logrank_test(data1, data2, alpha=95)
 
 
 def test_proportional_hazard_test():